remove unneeded query
